+++ packages/terra-consumer-nav/src/Nav.jsx
@@ -5,48 +5,20 @@ import Button from 'terra-button';
 import IconClose from 'terra-icon/lib/icon/IconClose';
 import NavItems from './components/nav-items/NavItems';
 import NavLogo from './components/nav-logo/NavLogo';
-import NavProfile from './components/nav-profile/NavProfile';
+import navItemShape from './NavPropShapes';
 import QuickLink from './components/quick-links/QuickLink';
 import QuickLinks from './components/quick-links/QuickLinks';
 import styles from './Nav.scss';
+import UserProfile from './components/user-profile/UserProfile';
 
 const cx = classNames.bind(styles);
-
-const navItemShape = {
-  /**
-   * The path the nav item should lead to.
-   */
-  uri: PropTypes.string.isRequired,
-  /**
-   * The text displayed on the link.
-   */
-  text: PropTypes.string.isRequired,
-  /**
-   * Whether or not the link should be styled as active or not.
-   */
-  isActive: PropTypes.bool,
-  /**
-   * An optional badge. When supplied, displays the value inline, styled alongside the text.
-   */
-  badgeValue: PropTypes.number,
-};
 
 const propTypes = {
   /**
    * An array of objects to be displayed as quick link options.
    */
-  quickLinks: PropTypes.arrayOf(
-    PropTypes.shape({
-      /**
-       * The path the nav item should lead to.
-       */
-      uri: PropTypes.string.isRequired,
-      /**
-       * The text displayed on the link.
-       */
-      text: PropTypes.string.isRequired,
-    }),
-  ),
+  quickLinks: PropTypes.arrayOf(PropTypes.shape(
+    navItemShape)),
   /**
    * An array of objects to be displayed as nav link options.
    */
@@ -61,6 +33,23 @@ const propTypes = {
       },
     ),
   ),
+  /**
+   * An array of nav items to be displayed on the user profile/ settings menu/popup.
+   */
+  profileLinks: PropTypes.arrayOf(PropTypes.shape(
+    navItemShape)),
+  /**
+   * User name to be displayed in the profile in navigation.
+   */
+  userName: PropTypes.string.isRequired,
+  /**
+   * Avatar to be displayed in user profile in navigation.
+   */
+  avatar: PropTypes.PropTypes.element,
+  /**
+   * The path signout button would redirect to.
+   */
+  signoutUrl: PropTypes.string.isRequired,
   /**
    * An object defining the logo to be displayed
    */
@@ -91,11 +80,13 @@ const propTypes = {
 const defaultProps = {
   quickLinks: [],
   navItems: [],
+  profileLinks: [],
+  avatar: null,
   logo: {},
 };
 
 const Nav = ({
-  quickLinks, navItems, logo, isMobileNavOpen, onRequestClose, ...customProps
+  quickLinks, navItems, profileLinks, userName, avatar, signoutUrl, logo, isMobileNavOpen, onRequestClose, ...customProps
 }) => (
   <div className={cx('nav-container')} {...customProps}>
     {/* Make this into a Slide Component */}
@@ -106,10 +97,11 @@ const Nav = ({
         {quickLinks.map(element => <QuickLink {...element} key={element.text} />)}
       </QuickLinks>
       <NavItems navItems={navItems} />
-      <NavProfile />
+      <UserProfile profileLinks={profileLinks} name={userName} avatar={avatar} id="profile-popup-button" signoutUrl={signoutUrl} />
     </div>
   </div>
 );
+
 
 Nav.propTypes = propTypes;
 Nav.defaultProps = defaultProps;