+++ packages/terra-consumer-nav/tests/nightwatch/Layout.jsx
@@ -2,24 +2,36 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import IconMenu from 'terra-icon/lib/icon/IconMenu';
+import { I18nProvider, i18nLoader } from 'terra-i18n';
+import Messages from '../../src/i18n/translations/messages.json';
 import Nav from '../../src/Nav';
+import NavHelp from '../../src/components/nav-help/NavHelp';
 import styles from './Layout.scss';
 
 const cx = classNames.bind(styles);
 
 const propTypes = {
   nav: PropTypes.object,
+  helpItems: PropTypes.array,
+  locale: PropTypes.oneOf(['en', 'en-GB', 'en-US', 'es']).isRequired,
 };
 
 class Layout extends React.Component {
-  constructor() {
-    super();
+  constructor(props) {
+    super(props);
 
     this.state = {
       isMobileNavOpen: false,
+      areTranslationsLoaded: false,
+      locale: props.locale,
+      messages: {},
     };
 
     this.toggleNav = this.toggleNav.bind(this);
+  }
+
+  componentDidMount() {
+    i18nLoader(this.props.locale, this.setState, this);
   }
 
   toggleNav() {
@@ -29,21 +41,31 @@ class Layout extends React.Component {
   }
 
   render() {
-    const { nav, ...customProps } = this.props;
+    const { nav, helpItems, locale, ...customProps } = this.props;
     return (
-      <div className={cx('layout', customProps.className)} {...customProps}>
-        <Nav
-          {...nav}
-          isMobileNavOpen={this.state.isMobileNavOpen}
-          onRequestClose={this.toggleNav}
-        />
-        <div className={cx('main-container')}>
-          <button className={cx('nav-burger')} onClick={this.toggleNav}>
-            <IconMenu />
-          </button>
-          <h2>I am in the main content</h2>
+      <I18nProvider
+        locale={this.state.locale}
+        messages={Object.assign({}, this.state.messages, Messages[this.state.locale])}
+      >
+        <div className={cx('layout', customProps.className)} {...customProps}>
+          <Nav
+            {...nav}
+            isMobileNavOpen={this.state.isMobileNavOpen}
+            onRequestClose={this.toggleNav}
+          />
+          <div className={cx('main-container')}>
+            <button className={cx('nav-burger')} onClick={this.toggleNav}>
+              <IconMenu />
+            </button>
+            {/* Added a div to test the HelpButton relative to page content */}
+            <div style={{ background: '#fff', height: '100%', width: 'inherit' }}>I am in the main content</div>
+            <div className={cx('footer')}>
+              <NavHelp helpNavs={helpItems} id="nav-help-button" />
+            </div>
+            <NavHelp className={cx('help-button-desktop')} helpNavs={helpItems} id="nav-help-button" />
+          </div>
         </div>
-      </div>
+      </I18nProvider>
     );
   }
 }